backend/solve-task/requirements.txt
openai==0.28.1
psycopg2-binary==2.9.9